Extract DiscogsCredentials from LibraryManager
The discogs API key and its validation state were mixed into the manager's
config surface. Move that concern into its own type holding clones of the two
handles those methods touch (the config handle and the key service); the
manager keeps one field and delegates its public discogs API to it.

discogs_validation_observer had no production caller — it was only ever an
internal helper of discogs_client and a test. Following the SyncController
pattern for test-only accessors, it lives on DiscogsCredentials and the test
reaches it through the field rather than a dead manager delegator.
