+++ bae-core/src/library/discogs_credentials.rs
@@ -1,0 +1,139 @@
+//! The Discogs-credentials responsibility extracted from [`LibraryManager`]: the
+//! stored API key (keyring-backed) and its validation state (config-backed),
+//! plus the client built from them.
+//!
+//! `LibraryManager` holds one `DiscogsCredentials` and delegates its public
+//! Discogs API to it. The type never references the manager back; it touches
+//! only the config handle (validation state) and the key service (the key
+//! itself).
+
+use std::sync::Arc;
+
+use crate::config::ConfigHandle;
+use crate::keys::{BaeKeyServiceExt, KeyService};
+
+/// Owns the Discogs API key and its validation state. Holds clones of the two
+/// handles those touch — the config handle (stored-key hint + validation) and
+/// the key service (the keyring-backed key). Cloned alongside the manager; both
+/// fields are themselves clone-shared handles.
+#[derive(Clone)]
+pub(crate) struct DiscogsCredentials {
+    config_handle: Arc<ConfigHandle>,
+    key_service: KeyService,
+}
+
+impl DiscogsCredentials {
+    pub(crate) fn new(config_handle: Arc<ConfigHandle>, key_service: KeyService) -> Self {
+        Self {
+            config_handle,
+            key_service,
+        }
+    }
+
+    pub(crate) fn has_discogs_token(&self) -> bool {
+        self.config_handle.has_discogs_key()
+    }
+
+    pub(crate) fn get_discogs_token(&self) -> Result<Option<String>, String> {
+        self.key_service
+            .get_discogs_key()
+            .map_err(|e| e.to_string())
+    }
+
+    pub(crate) fn save_discogs_key(&self, token: &str) -> Result<(), String> {
+        self.key_service
+            .set_discogs_key(token)
+            .map_err(|e| e.to_string())
+    }
+
+    pub(crate) fn delete_discogs_key(&self) -> Result<(), String> {
+        self.key_service
+            .delete_discogs_key()
+            .map_err(|e| e.to_string())
+    }
+
+    /// Record a stored key with its validation state — the single write for the
+    /// save path. `Some(validation)` is both the stored-key hint and the state,
+    /// so one `update` keeps them consistent and fires one watch-channel
+    /// notification.
+    pub(crate) fn set_discogs_key_stored(
+        &self,
+        validation: crate::config::DiscogsValidation,
+    ) -> Result<(), crate::config::ConfigError> {
+        self.config_handle.update(|c| c.discogs = Some(validation))
+    }
+
+    /// Clear the stored-key state — no key, so no validation.
+    pub(crate) fn clear_discogs_key_stored(&self) -> Result<(), crate::config::ConfigError> {
+        self.config_handle.update(|c| c.discogs = None)
+    }
+
+    /// Update the stored key's validation state. No-op when no key is stored —
+    /// validation describes a key that exists.
+    pub(crate) fn set_discogs_validation(
+        &self,
+        validation: crate::config::DiscogsValidation,
+    ) -> Result<(), crate::config::ConfigError> {
+        self.config_handle.update(|c| {
+            if c.discogs.is_some() {
+                c.discogs = Some(validation);
+            }
+        })
+    }
+
+    pub(crate) fn discogs_validation(&self) -> Option<crate::config::DiscogsValidation> {
+        self.config_handle.config().discogs
+    }
+
+    /// An observer that folds a Discogs call's outcome into the stored key's
+    /// validation, so every call site updates the stored validation state
+    /// without recording the outcome itself. A 401 marks it `Rejected`; a
+    /// success while it was `Unvalidated` confirms it `Valid`; any other outcome
+    /// (network, rate-limit, success while already settled, or no key stored)
+    /// leaves it untouched. Injected into the client by [`Self::discogs_client`].
+    pub(crate) fn discogs_validation_observer(
+        &self,
+    ) -> crate::discogs::client::DiscogsValidationObserver {
+        use crate::config::DiscogsValidation;
+        use crate::discogs::client::DiscogsKeySignal;
+        let config_handle = self.config_handle.clone();
+        std::sync::Arc::new(move |signal| {
+            let Some(current) = config_handle.config().discogs else {
+                // The key was removed while a Discogs call was in flight: the
+                // client outlives the config entry it was built from. Nothing to
+                // fold the outcome into.
+                tracing::debug!("discogs validation signal ignored: no key stored");
+                return;
+            };
+            let next = match signal {
+                DiscogsKeySignal::Rejected => DiscogsValidation::Rejected,
+                DiscogsKeySignal::Accepted if current == DiscogsValidation::Unvalidated => {
+                    DiscogsValidation::Valid
+                }
+                _ => return,
+            };
+            if current == next {
+                return;
+            }
+            if let Err(e) = config_handle.update(|c| c.discogs = Some(next)) {
+                tracing::warn!("failed to persist discogs validation {next:?}: {e}");
+            }
+        })
+    }
+
+    /// A client for the stored key, unless that key is `Rejected`. A `Valid` or
+    /// `Unvalidated` key is served (the latter used optimistically); a
+    /// `Rejected` key is withheld so search call sites skip Discogs entirely.
+    /// The client reports each call's outcome back into the validation state.
+    pub(crate) fn discogs_client(&self) -> Result<Option<crate::discogs::DiscogsClient>, String> {
+        if self.discogs_validation() == Some(crate::config::DiscogsValidation::Rejected) {
+            return Ok(None);
+        }
+        let observer = self.discogs_validation_observer();
+        Ok(self
+            .key_service
+            .get_discogs_key()
+            .map_err(|e| e.to_string())?
+            .map(|key| crate::discogs::DiscogsClient::with_observer(key, observer)))
+    }
+}
