+++ bae-core/src/library/manager/config.rs
@@ -88,25 +88,19 @@ impl LibraryManager {
     // =========================================================================
 
     pub fn has_discogs_token(&self) -> bool {
-        self.config_handle.has_discogs_key()
+        self.discogs.has_discogs_token()
     }
 
     pub fn get_discogs_token(&self) -> Result<Option<String>, String> {
-        self.key_service
-            .get_discogs_key()
-            .map_err(|e| e.to_string())
+        self.discogs.get_discogs_token()
     }
 
     pub fn save_discogs_key(&self, token: &str) -> Result<(), String> {
-        self.key_service
-            .set_discogs_key(token)
-            .map_err(|e| e.to_string())
+        self.discogs.save_discogs_key(token)
     }
 
     pub fn delete_discogs_key(&self) -> Result<(), String> {
-        self.key_service
-            .delete_discogs_key()
-            .map_err(|e| e.to_string())
+        self.discogs.delete_discogs_key()
     }
 
     /// Record a stored key with its validation state — the single write for the
@@ -117,12 +111,12 @@ impl LibraryManager {
         &self,
         validation: crate::config::DiscogsValidation,
     ) -> Result<(), crate::config::ConfigError> {
-        self.config_handle.update(|c| c.discogs = Some(validation))
+        self.discogs.set_discogs_key_stored(validation)
     }
 
     /// Clear the stored-key state — no key, so no validation.
     pub fn clear_discogs_key_stored(&self) -> Result<(), crate::config::ConfigError> {
-        self.config_handle.update(|c| c.discogs = None)
+        self.discogs.clear_discogs_key_stored()
     }
 
     /// Update the stored key's validation state. No-op when no key is stored —
@@ -131,51 +125,11 @@ impl LibraryManager {
         &self,
         validation: crate::config::DiscogsValidation,
     ) -> Result<(), crate::config::ConfigError> {
-        self.config_handle.update(|c| {
-            if c.discogs.is_some() {
-                c.discogs = Some(validation);
-            }
-        })
+        self.discogs.set_discogs_validation(validation)
     }
 
     pub fn discogs_validation(&self) -> Option<crate::config::DiscogsValidation> {
-        self.config_handle.config().discogs
-    }
-
-    /// An observer that folds a Discogs call's outcome into the stored key's
-    /// validation, so every call site updates the stored validation state
-    /// without recording the outcome itself. A 401 marks it `Rejected`; a
-    /// success while it was `Unvalidated` confirms it `Valid`; any other outcome
-    /// (network, rate-limit, success while already settled, or no key stored)
-    /// leaves it untouched. Injected into the client by [`Self::discogs_client`].
-    pub(crate) fn discogs_validation_observer(
-        &self,
-    ) -> crate::discogs::client::DiscogsValidationObserver {
-        use crate::config::DiscogsValidation;
-        use crate::discogs::client::DiscogsKeySignal;
-        let config_handle = self.config_handle.clone();
-        std::sync::Arc::new(move |signal| {
-            let Some(current) = config_handle.config().discogs else {
-                // The key was removed while a Discogs call was in flight: the
-                // client outlives the config entry it was built from. Nothing to
-                // fold the outcome into.
-                tracing::debug!("discogs validation signal ignored: no key stored");
-                return;
-            };
-            let next = match signal {
-                DiscogsKeySignal::Rejected => DiscogsValidation::Rejected,
-                DiscogsKeySignal::Accepted if current == DiscogsValidation::Unvalidated => {
-                    DiscogsValidation::Valid
-                }
-                _ => return,
-            };
-            if current == next {
-                return;
-            }
-            if let Err(e) = config_handle.update(|c| c.discogs = Some(next)) {
-                tracing::warn!("failed to persist discogs validation {next:?}: {e}");
-            }
-        })
+        self.discogs.discogs_validation()
     }
 
     /// A client for the stored key, unless that key is `Rejected`. A `Valid` or
@@ -183,14 +137,6 @@ impl LibraryManager {
     /// `Rejected` key is withheld so search call sites skip Discogs entirely.
     /// The client reports each call's outcome back into the validation state.
     pub fn discogs_client(&self) -> Result<Option<crate::discogs::DiscogsClient>, String> {
-        if self.discogs_validation() == Some(crate::config::DiscogsValidation::Rejected) {
-            return Ok(None);
-        }
-        let observer = self.discogs_validation_observer();
-        Ok(self
-            .key_service
-            .get_discogs_key()
-            .map_err(|e| e.to_string())?
-            .map(|key| crate::discogs::DiscogsClient::with_observer(key, observer)))
+        self.discogs.discogs_client()
     }
 }
