+++ bae-core/src/library/manager/mod.rs
@@ -42,6 +42,7 @@ use crate::db::{
 };
 use crate::keys::BaeKeyServiceExt;
 use crate::keys::KeyService;
+use crate::library::discogs_credentials::DiscogsCredentials;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
 use crate::library::export::ExportService;
 use crate::library::sync_controller::SyncController;
@@ -568,6 +569,10 @@ pub struct LibraryManager {
     /// encryption-service cell, membership, and the coven make-Remote/make-Local
     /// primitives. The manager delegates its public sync API here.
     sync: SyncController,
+    /// The Discogs-credentials responsibility: the stored API key (keyring) and
+    /// its validation state (config), plus the client built from them. The
+    /// manager delegates its public Discogs API here.
+    discogs: DiscogsCredentials,
     /// Cancellation tokens for in-progress foreground transfers (unmanage),
     /// keyed by release id. `cancel_release_transition` fires the token; the
     /// transfer observes it between files, deletes the partial copies it wrote,
@@ -619,6 +624,7 @@ impl Clone for LibraryManager {
             handle: self.handle.clone(),
             event_tx: self.event_tx.clone(),
             sync: self.sync.clone(),
+            discogs: self.discogs.clone(),
             transfer_cancels: self.transfer_cancels.clone(),
             download_queue: self.download_queue.clone(),
             #[cfg(any(test, feature = "test-utils"))]
@@ -681,6 +687,8 @@ impl LibraryManager {
             sync_connected,
         );
 
+        let discogs = DiscogsCredentials::new(config_handle.clone(), key_service.clone());
+
         Ok(LibraryManager {
             database,
             library_dir,
@@ -692,6 +700,7 @@ impl LibraryManager {
             handle,
             event_tx,
             sync,
+            discogs,
             transfer_cancels: Arc::new(Mutex::new(HashMap::new())),
             download_queue: Arc::new(crate::library::DownloadQueue::new()),
             #[cfg(any(test, feature = "test-utils"))]
@@ -729,6 +738,8 @@ impl LibraryManager {
             Arc::new(std::sync::atomic::AtomicBool::new(false)),
         );
 
+        let discogs = DiscogsCredentials::new(config_handle.clone(), key_service.clone());
+
         LibraryManager {
             database,
             library_dir,
@@ -740,6 +751,7 @@ impl LibraryManager {
             handle,
             event_tx,
             sync,
+            discogs,
             transfer_cancels: Arc::new(Mutex::new(HashMap::new())),
             download_queue: Arc::new(crate::library::DownloadQueue::new()),
             #[cfg(any(test, feature = "test-utils"))]
