+++ bae-core/src/library/mod.rs
@@ -1,4 +1,5 @@
 pub mod app_services;
+mod discogs_credentials;
 pub mod download_queue;
 pub mod download_snapshot;
 #[cfg(not(any(target_os = "ios", target_os = "android")))]
